requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
dj-rest-auth==5.0.2
Django==5.0.6
django-allauth==0.62.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
idna==3.7
mysqlclient==2.2.4
packaging==24.1
pycparser==2.22
PyJWT==2.8.0
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.2
